Refactored projects/tasks.py
