+++ readthedocs/rtd_tests/tests/test_celery.py
@@ -10,6 +10,7 @@ from mock import MagicMock, patch
 
 from readthedocs.builds.constants import LATEST
 from readthedocs.builds.models import Build
+from readthedocs.builds.tasks import fileify
 from readthedocs.doc_builder.exceptions import VersionLockedError
 from readthedocs.projects import tasks
 from readthedocs.builds.models import Version
@@ -267,5 +268,5 @@ class TestCeleryBuilding(RTDTestCase):
     @patch('readthedocs.builds.managers.log')
     def test_fileify_logging_when_wrong_version_pk(self, mock_logger):
         self.assertFalse(Version.objects.filter(pk=345343).exists())
-        tasks.fileify(version_pk=345343, commit=None)
+        fileify(version_pk=345343, commit=None)
         mock_logger.warning.assert_called_with("Version not found for given kwargs. {'pk': 345343}")
