+++ readthedocs/rtd_tests/tests/test_project.py
@@ -16,9 +16,9 @@ from readthedocs.builds.constants import (
     LATEST,
 )
 from readthedocs.builds.models import Build
+from readthedocs.builds.tasks import finish_inactive_builds
 from readthedocs.projects.exceptions import ProjectConfigurationError
 from readthedocs.projects.models import Project
-from readthedocs.projects.tasks import finish_inactive_builds
 from readthedocs.rtd_tests.mocks.paths import fake_paths_by_regex
